:sparkles: Add support for making a day as done
This is far from the final form of this -- this doesn't even really work in
that I can't sensibly store the done information in the widgets really as
they get shuffled about and destroyed. This is just a proof-of-concept of
how the display will work and how it'll be styled, etc.

The real approach will be to move the data up into the streak container,
most likely.