:hammer: Move the edit logic out of the screen and into the StreakLine
For a while here I thought I wanted to keep as much of the logic as possible
in the Main screen. Thing is, I was starting to get unhelpful bindings
creeping in. That is, the binding for editing a streak's title was always
there even if no streaks had been added yet. Not ideal.

Now, of course, I could have made sure the code ignored any request to edit
a streak that wasn't there but... nah, that's not tidy. That's a code smell
for Textual that tells me I should move the binding to the widget involved.

So here I am doing that.

I think I still want to keep some stuff on the Main screen. Adding a streak
makes sense here. Also, any styling to do with colour and the like... I
think that generally does need to stay out of the streak widget. I want the
screen to be responsible for how the widgets look and where they go, but I
think it's fine for the widgets to be fully responsible for their own
working.