:fire: Undo the animation of background change
This causes a drawing problem with zoom, mainly I think down to a confusion
as to what rules are applying from CSS due to the heavy use of DEFAULT_CSS.
Part of the point of this particular app was to see where this would take
me, not using an external CSS file. I think some more thought needs to go
into this.