:zap: Make the cursor keys work again
Due to binding inheritance being introduced in 0.6.0, the cursor keys
stopped working in the app. The issue is that they're getting consumed by
widgets *below* the screen now. This makes the screen as having priority for
those keys.

See https://github.com/Textualize/textual/issues/1343 for a bunch of
background on this, and note that this will change some more in the future
once the fix we're aiming for goes into place.