:hammer: Get StreakLine removal working again
The upgrade to 0.14 of Textual meant the old streak line removal stopped
working (honestly, I was always a little surprised that the way I had it
working actually worked in the first place).

This approach doesn't work as I'd like, and in digging a little bit it might
have uncovered a bug in Textual regarding posting messages and widget
removal, see this issue for context:

    https://github.com/Textualize/textual/issues/2017

But... the approach I've taken here does work. It's more or less what I want
anyway, I just delay the actual act of removal.